==> notebooks/01-EDAofComentsFromUKRvideos.ipynb <==
import pandas as pd

path = '../data/comments.csv'

df = pd.read_csv(path)

df.head()

len(df)

df.columns[3]

df.groupby(['video_id']).count().sort_values(by=['id'], ascending = False)

df[df.author == 'Олежка Русский']['video_id'].nunique()

x = dict()
for person in df.author.unique():
    x[person] =df[df.author == person]['video_id'].nunique()

import operator
sorted_x = sorted(x.items(), key=operator.itemgetter(1), reverse=True)

sorted_x

for text in df[df.author == 'Иван Иванов'].text:
    print(text)
    print('-----------------------')

import nltk

nltk.download('punkt')

nltk.download('averaged_perceptron_tagger')

sentence =  df[(df.author == 'Иван Иванов') & (df.id =='Ugi6ohMVoexb1ngCoAEC')].text.values[0]

sentence

tokens = nltk.word_tokenize(sentence)

tokens

tagged = nltk.pos_tag(tokens)

tagged

from nltk.stem.snowball import SnowballStemmer
stemmer = SnowballStemmer("russian") 

stemmer.stem("БЛЯДЬ")

bad_words = open('../data/bad_words.txt').read()

bad_words = bad_words.split(', ')

bad_words

count = 0
error_count =0 
for sentence in df.text:
    try:
        tokens = nltk.word_tokenize(sentence)
    except TypeError:
        error_count += 1
    if any(bad_word in tokens for bad_word in bad_words):
        count += 1

print(count)

error_count

len(df)
