==> notebooks/03-Stemming.ipynb <==
import pandas as pd
import nltk

df = pd.read_csv('../data/comments_clean1.csv')

bad_words = open('../data/bad_words_seed.txt').read().split('\n')[:-1]

df = df.dropna()

from nltk.stem.snowball import SnowballStemmer
stemmer = SnowballStemmer("russian") 

bad_words_stemmed = []

for word in bad_words:
    bad_words_stemmed.append(stemmer.stem(word))

bad_words_stemmed = [i for i in bad_words_stemmed if len(i) > 2]

bad_words_stemmed = list(set(bad_words_stemmed))

bad_words_stemmed.append('хуй')
bad_words_stemmed.append('хуе')

def lable_stammed(x):
    tokens = nltk.word_tokenize(x)
    #tokens = [stemmer.stem(token) for token in tokens]
    for bad_word in bad_words_stemmed:
        for token in tokens:
            if bad_word in token:
                return True
    return False    

def lable(x):
    tokens = nltk.word_tokenize(x)
    if any(bad_word in tokens for bad_word in bad_words):
        return True
    return False 

df['label'] = df.text.apply(lable)

df.head()

df.to_csv('../data/lable.csv', index = False)

df['label'] = df.text.apply(lable_stammed)

df.to_csv('../data/lable_stammed.csv', index = False)
