==> notebooks/08-ngrams.ipynb <==
import pandas
import operator
import numpy as np
from nltk import ngrams
import seaborn as sns
import nltk
from nltk.stem.snowball import SnowballStemmer
%matplotlib inline

df = pandas.read_csv('../data/lable_stammed.csv')

len(df[df.label == True])

stemmer = SnowballStemmer("russian", ignore_stopwords=True) 

sentence = str(df.sample(1).text.values[0])

sentence = "Ebat"

bigrams = ngrams(sentence.split(), 2)
for grams in bigrams:
    print(grams)
    print(stemmer.stem(grams[0]), stemmer.stem(grams[1]))

all_grams = dict()
for sentence in df.text:
    bigrams = ngrams(sentence.split(), 2)
    for grams in bigrams:
        key = str(stemmer.stem(grams[0])) + str(stemmer.stem(grams[1]))
        if key in all_grams:
            all_grams[key] += 1
        else:
            all_grams[key] = 1
