==> notebooks/09-RandomSample.ipynb <==
import pandas as pd
import copy

df = pd.read_csv('../data/lable_stammed.csv')


df.dropna()

manual = pd.read_csv('../data/manual.csv')

len(manual)

len(df[df.label == True])

save = copy.copy(sample)

sample = df.sample(1000)

len(sample[sample.label == True])

sample['hate'] =False

sample.head()

sample['evaluated'] = False

subsample = sample[(sample.label == False) & (sample.hate == False) & (sample.evaluated == False)]

count = 0
for i in subsample.index:
    print(count)
    print(sample.loc[i]['text'])
    label = int(input('Abusive? :'))
    if label == 1:
        sample.at[i, 'label'] = True
    elif label == 2:
        sample.at[i, 'label'] = True
        sample.at[i, 'hate'] = True    
    elif label == 0:
        break
    sample.at[i, 'evaluated'] = True
    count += 1
    
    

sample[(sample.label == False) & (sample.hate == False)].sample(1).text.values

def mark(x):
    if any(b in x for b in ['хохл','кацап','майдаун','майданут','бендер', 'западенские', 'москал', 'жид', 'рашк']):
        return True
    else:
        return False


sample['hate'] = sample.text.apply(mark)

len(sample[(sample.label == True) | (sample.hate == True)])

for i in range(len(sample)):
    if sample.iat[i, 6] == True:
        sample.iat[i, 5] == True

sample.head()

subsample.head()

sample.at[7334, 'text']

len(sample[(sample.label == False) & (sample.hate == False) & (sample.evaluated == False)])

len(sample[(sample.label == True)])

len(sample[(sample.hate == True)])

sample.to_csv('../data/manual.csv', index = False)
