==> notebooks/10-BadWordStemming.ipynb <==
import nltk
from nltk.stem.snowball import SnowballStemmer

stemmer = SnowballStemmer("russian", ignore_stopwords=True) 

bad_words = open('../data/bad_words.txt').read().split('\n')[:-1]

stemmed_dictionary = []
for word in bad_words:
    stemmed_dictionary.append(stemmer.stem(word))

stemmed_dictionary

stemmer.stem('хохлятских')

if 'хуй' in 'хуй':
    print('aaaa')
