==> notebooks/12-Pipeline.ipynb <==
import pandas as pd
from alphabet_detector import AlphabetDetector
import string
import re
import nltk
import numpy as np
import operator
from nltk.stem.snowball import SnowballStemmer

# Data cleaning

path_to_data = '../data/comments1.csv'

df = pd.read_csv(path_to_data)
df = df.dropna()

emoji_pattern = re.compile("["
        u"\U0001F600-\U0001F64F"  # emoticons
        u"\U0001F300-\U0001F5FF"  # symbols & pictographs
        u"\U0001F680-\U0001F6FF"  # transport & map symbols
        u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                           "]+", flags=re.UNICODE)

def clean(x):
    ad = AlphabetDetector()
    res = x
    for ch in string.punctuation:                                                                                                     
        res = res.replace(ch, ' ')
    res = ''.join([i for i in res if not i.isdigit()])
    res = res.lower()
    res = emoji_pattern.sub(r' ', res)
    res = res.replace('\n', ' ')
    res = res.replace('\t', ' ')
    res = res.replace('\ufeff', ' ')
    res = res.replace('\r\n', '  ')
    res = res.replace('\xa0', ' ')
    res = res.replace('ё', 'е')
    res = re.sub(' +',' ', res)
    if  not ad.only_alphabet_chars(res, "CYRILLIC"): 
        res = ''
    return res

df.text = df.text.apply(clean)
df = df[df.text != '']

#df.to_csv('../data/comments_clean1.csv', index=False)

#remove comments that were manualy labeld
manual = pd.read_csv('../data/manual.csv')
df = df[~df.id.isin(manual.id)]

# Stemming bad words

def stem(word):
    stemmer = SnowballStemmer("russian", ignore_stopwords=True) 
    stemmed_word = stemmer.stem(word)
    if len(stemmed_word) <= 2:
        return word
    return stemmed_word

path_to_bad_words = '../data/bad_words.txt'

bad_words = open(path_to_bad_words).read().split('\n')[:-1]

report_dictionary = [] # array that will contain all bad words
stemmed_dictionary = [] # array that will contain actual stems used for finding bad words

for word in bad_words:
    report_dictionary.append(word)
    stemmed_dictionary.append(stem(word))

stemmed_dictionary = [i for i in stemmed_dictionary if len(i) > 2] 

stemmed_dictionary = list(set(stemmed_dictionary))

stemmed_dictionary.append('хуй') # nltk stemmer can not corectly stem word хуй
stemmed_dictionary.append('хуе')

# label coments with bad words

def label(x):
    global stemmed_dictionary
    tokens = nltk.word_tokenize(x)
    for bad_word in stemmed_dictionary:
        for token in tokens:
            if bad_word in token:
                return True
    return False

df['label'] = df.text.apply(label)

# Iterative process of finding new bad words

## Word stemming and counting

all_stemmed = dict()
for sentence in df.text:
    for token in nltk.word_tokenize(sentence):
        if token not in all_stemmed:
            all_stemmed[token] = stem(token)

all_counts = dict()
for sentence in df.text:
    tokens = nltk.word_tokenize(sentence)
    for token in tokens:
        word = stem(token)
        if word in all_counts:
            all_counts[word] += 1
        else:
            all_counts[word] = 1

# Likelihood of words being in comments labeld as bad or labeld as not bad

def likelihood(label):
    labeld_part = dict()
    part = df[df.label == label]
    for index, row in part.iterrows():
        sentence = row.text
        tokens = nltk.word_tokenize(sentence)
        for token in tokens:
            stemed = stem(token)
            if stemed != '':
                if all_counts[stemed] >= 20:
                    if stemed in labeld_part:
                        labeld_part[stemed] += 1.0/len(part)
                    else:
                        labeld_part[stemed] = 1.0/len(part)
    return labeld_part

## Criteria functions for decision if word is bad or not bad

def relative_distance(p_good, p_bad):
    difference = dict()
    for key in p_good:
        if key in p_bad:
            difference[key] =(p_bad[key] - p_good[key])/ np.maximum( p_bad[key], p_good[key]) 
    return difference

def logg_odds(p_good, p_bad):
    ratio = dict()
    for key in p_good:
        if key in p_bad:
            odds_good =  p_good[key]/(1 -  p_good[key])
            odds_bad = p_bad[key]/(1- p_bad[key])
            ratio[key] = np.log(odds_bad/odds_good)
    return ratio

## Pointwise mutual information

def make_pairs(df_bad):
    pairs = dict()
    for sentence in df_bad.text:
        selection = []
        for token in nltk.word_tokenize(sentence):
            word = stem(token)
            if word in stemmed_dictionary and all_counts[word] >= 20:
                selection.append(word)
        for bad_word in selection:
            for token in nltk.word_tokenize(sentence):
                word = stem(token)
                if bad_word != word and word != '' and all_counts[word] >= 20:
                    key = bad_word + '-' + word
                    if key in pairs:
                        pairs[key] += 1
                    else:
                        pairs[key] = 1
    return pairs

def pmi():
    bad_likelihood = likelihood(True)
    pairs = make_pairs(df[df.label == True])
    result = dict()
    for key in pairs:
        words = key.split('-')
        pxy = float(pairs[key]/len(df[df.label == True]))
        px = bad_likelihood[words[0]]
        py = bad_likelihood[words[1]]
        result[key] = np.log(pxy/(px*py))
    return result
        
        
        

sort(pmi())

## Detecting new bad words

def unstem(stem):
    res = []
    for key in all_stemmed:
        if all_stemmed[key] == stem:
            res.append(key)
    return res    

def sort(x, rev = True):
    return sorted(x.items(), key=operator.itemgetter(1), reverse=rev)

def update_dictionary(word):
    global stemmed_dictionary
    if len(word) <= 2:
        return False
    #for key in stemmed_dictionary:
        #if word in key:
            #return False
    stemmed_dictionary.append(word)
    return True

def iterate(df, report_dictionary, stemmed_dictionary):
    iter = True
    while(iter):
        print('-----------------------------------')
        print('labeld as abusive: ', len(df[df.label == True]))
        #print('List of bad_words')
        #print(report_dictionary)
        p_bad = likelihood(True)
        p_good = likelihood(False)
        prob = logg_odds(p_good, p_bad)
        print(sort(prob)[:10])
        print('New bad words')
        new_words = []
        for key in prob:
            if prob[key] > 3:
                if update_dictionary(key):
                    new_words += unstem(key)
                    print( unstem(key))
        report_dictionary += new_words
        if len(new_words) == 0:
            iter = False
            break
        #print('\nNew dictionary')
        #print(report_dictionary)
        df['label'] =  df.text.apply(label)
        print('relabeld as abusive: ', len(df[df.label == True]))

iterate(df, report_dictionary, stemmed_dictionary)

# Evaluation

manual['evaluation'] = manual.text.apply(label)

tp = 0
tn = 0
fp = 0
fn = 0
for index, row in manual.iterrows():
    if row.label == True and row.evaluation == True:
        tp += 1
    if row.label == False and row.evaluation == False:
        tn += 1
    if row.label == False and row.evaluation == True:
        fp += 1
    if row.label == True and row.evaluation == False:
        #print(row.text)
        fn += 1
accuracy = (tp + fp)/(tp + fp + fn + fp)
precision = tp/(tp + fp)
recall = tp/(tp + fn)
f1  = 2 * (precision * recall)/(precision + recall)
print('accuracy: ', accuracy)
print('precision: ', precision)
print('recall: ', recall)
print('f1: ', f1)
