==> notebooks/13-ManualLabeling.ipynb <==
import pandas as pd

manual = pd.read_csv('../data/manual.csv')

df = pd.read_csv('../data/comments_clean1.csv')

df = df[~df.id.isin(manual.id)]

sample = df.sample(1000)

sample['label'] = False 

sample['hate'] = False

sample['evaluated'] = False #This varaible checks if coment was manualy labeled

# this subsample is used for the case if you want to stop labeling at some moment and then continue
subsample = sample[(sample.label == False) & (sample.hate == False) & (sample.evaluated == False)]

# Press 1 if coment has abusive language
# Press 2 if coment has hateful language (e.g хахалы)
# Press 3 if coment doesn't any abusive and hateful language
# Press 0 if you want to stop 
count = 0
for i in subsample.index:
    print(count)
    print(sample.loc[i]['text'])
    label = int(input('Abusive? :'))
    if label == 1:
        sample.at[i, 'label'] = True # Coment has abusive language
    elif label == 2:
        sample.at[i, 'hate'] = True   # Coment has hateful language
        sample.at[i, 'label'] = True  # All hateful language is abusive by default    
    elif label == 0:
        break
    sample.at[i, 'evaluated'] = True # Set that comment was manualy labeled
    count += 1

if len(sample[sample.evaluated == False]) == 0:
    new_labeled = pd.concat([manual, sample])
    new_labeled.to_csv('../data/manual.csv', index=False)
else:
    print('Please, label all coments in sample')
