==> notebooks/15-LemmatizationWithCorpus.ipynb <==
import pandas as pd
from alphabet_detector import AlphabetDetector
import string
import re
import nltk
import numpy as np
import operator
from nltk.stem.snowball import SnowballStemmer
import xml.etree.ElementTree

path_to_data = '../data/comments_clean1.csv'

df = pd.read_csv(path_to_data)

corpus = xml.etree.ElementTree.parse('../corpus/opcorpora.xml').getroot()

def lemmatize(word):
    for lemma in corpus.iter('lemma'):
        found = False
        for forms in lemma.getchildren():
            if forms.attrib['t'] == word:
                found = True
                break
        if found:
            for forms in lemma.getchildren():
                if forms.tag == 'l':
                    return forms.attrib['t']
    return False

lemmatize("ершей")

words = dict()
for sentence in df.text:
    tokens = nltk.word_tokenize(sentence)
    for token in tokens:
        if token not in words:
            words[token] = 0
            
    

for word in words:
    if lemmatize(word) == False:
        print(word)
