==> notebooks/16-WordsPerVideo.ipynb <==
import pandas as pd
from alphabet_detector import AlphabetDetector
import string
import re
import nltk
import numpy as np
import operator
from nltk.stem.snowball import SnowballStemmer
import xml.etree.ElementTree

df = pd.read_csv('../data/comments_clean1.csv')

df.head()

def stem(word):
    stemmer = SnowballStemmer("russian", ignore_stopwords=True) 
    stemmed_word = stemmer.stem(word)
    if len(stemmed_word) <= 2:
        return word
    return stemmed_word

all_counts = dict()
for sentence in df.text:
    tokens = nltk.word_tokenize(sentence)
    for token in tokens:
        word = stem(token)
        if word in all_counts:
            all_counts[word] += 1
        else:
            all_counts[word] = 1

len(all_counts)

word_to_id = dict()
for index, row in df.iterrows():
    tokens = nltk.word_tokenize(row.text)
    for token in tokens:
        word = stem(token)
        if word in word_to_id:
            word_to_id[word].append(row.video_id)
        else:
            word_to_id[word] = [row.video_id]

len(word_to_id)

for key in word_to_id:
    word_to_id[key] = list(set(word_to_id[key]))

word_to_id_count = dict()
for key in word_to_id:
    word_to_id_count[key] = len(word_to_id[key])

all_counts['иностранц']

word_to_id_count['иностранц']

count = 0
for key in word_to_id_count:
    if word_to_id_count[key] >2:
        count += 1


print(count)

df.video_id.value_counts()
